requirements.txt
# WDMPA - PC Dependencies (CUDA 11.x / RTX 4090)
# Install with: pip install -r requirements.txt

# Core
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Training
tqdm>=4.65.0
pandas>=2.0.0
matplotlib>=3.7.0
PyYAML>=6.0

# Model Analysis
thop>=0.1.1  # FLOPs counting
onnx>=1.14.0
onnxruntime-gpu>=1.15.0  # For ONNX validation

# Logging
tensorboard>=2.13.0
